20180527  修复了从托盘图标打开主界面时不会自动刷新的BUG  修复了在事件列表刷新时会出错的BUG  新增区别开机自启时间类型  修复了设置界面每页项目数量显示有小数点的BUG并将最大值增加到1000  新增记录用户、系统的10+个事件  新增F5刷新界面

+++ ComputerUsage/UserControls/UcSettings.xaml.cs
@@ -25,7 +25,7 @@ namespace ComputerUsage
         {
             InitializeComponent();
             sldInterval.TextConvert = p => p.ToString() + "秒";
-            sldItemsCountOfEachPage.TextConvert = p => p.ToString() + "条";
+            sldItemsCountOfEachPage.TextConvert = p => ((int)p).ToString() + "条";
             if(WpfCodes.Program.Startup.WillRunWhenStartup("ComputerUsage"))
             {
                 if(Set.NoWindowWhenStartup)
